migration of mod to api.v2 stage 2 (transport)

+++ source-platforms/forge/src/main/java/buildcraft/lib/misc/LocaleUtil.java
@@ -164,7 +164,7 @@ public class LocaleUtil {
         }
     }
 
-    public static MutableComponent localizeFluidFlow(int milliBucketsPerTick) {
+    public static MutableComponent localizeFluidFlow(long milliBucketsPerTick) {
         if (BCLibConfig.hideFluidValues) {
             return hiddenValue();
         }
